fix: add named component export
Closes: #10465
