+++ src/types/form.ts
@@ -845,3 +845,39 @@ export type FormProviderProps<
 > = {
   children: React.ReactNode | React.ReactNode[];
 } & UseFormReturn<TFieldValues, TContext, TTransformedValues>;
+
+export type FormProps<
+  TFieldValues extends FieldValues,
+  TTransformedValues extends FieldValues | undefined = undefined,
+> = Omit<React.FormHTMLAttributes<HTMLFormElement>, 'onError'> &
+  Partial<{
+    control: Control<TFieldValues>;
+    headers: Record<string, string>;
+    validateStatus: (status: number) => boolean;
+    onError: ({
+      response,
+      error,
+    }:
+      | {
+          response: Response;
+          error?: undefined;
+        }
+      | {
+          response?: undefined;
+          error: unknown;
+        }) => void;
+    onSuccess: ({ response }: { response: Response }) => void;
+    onSubmit: TTransformedValues extends FieldValues
+      ? FormSubmitHandler<TTransformedValues>
+      : FormSubmitHandler<TFieldValues>;
+    method: 'post' | 'put' | 'delete';
+    children: React.ReactNode | React.ReactNode[];
+    render: (props: {
+      submit: (e?: React.FormEvent) => void;
+    }) => React.ReactNode | React.ReactNode[];
+    encType:
+      | 'application/x-www-form-urlencoded'
+      | 'multipart/form-data'
+      | 'text/plain'
+      | 'application/json';
+  }>;
